# api/openapi.yaml
openapi: 3.1.0
info:
  title: Snooze API
  version: 2.0.0
  summary: Log aggregation and alerting REST API.
  description: |
    The Snooze v1 HTTP API. Snooze 2.0 is a Go rewrite of the Python 1.x
    server; this document describes the **v1** wire contract that ships
    with 2.0.

    ## Authentication

    Every endpoint outside the small "public" set (`/healthz`,
    `/readyz`, `/metrics`, `/api/v1/login/*`) requires a bearer JWT:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via `POST /api/v1/login/{local,ldap,anonymous}`.
    The Python 1.x `Authorization: JWT <token>` scheme is no longer
    accepted.

    ## Multi-tenancy

    Snooze supports multiple isolated organizations (tenants) on a single
    server. Every authenticated request is scoped to the tenant encoded in
    the `tenant_id` JWT claim. Pass the `org` field in the login request
    body to obtain a token scoped to a specific tenant; omit `org` (or leave
    it blank) to land in the `default` tenant.

    **Ingest routing** — unauthenticated alert ingestion (`POST /api/v1/alerts`,
    `POST /api/v1/webhook/*`) is routed to a tenant by the `Authorization:
    Bearer <ingest_token>` header, where `<ingest_token>` is the per-tenant
    opaque token stored in the tenant document. When no token is supplied or
    the token is unknown, the request falls back to the `default` tenant.

    **Platform permissions** — the `ro_tenant` and `rw_tenant` permissions gate
...

    ## CRUD shape
...
servers:
...
security:
  - bearerAuth: []
  # Trusted-header (auth-proxy) mode is an alternative to a Bearer token on
  # every authenticated route, but only when auth_proxy.enabled=true and the
  # immediate TCP peer is a trusted proxy. Disabled by default.
  - proxyHeader: []
...
paths:
...
  /api/v1/avatar/{method}/{name}:
    get:
      tags: [auth]
      summary: Get a user's profile picture
      description: |
        Returns the stored profile picture of the user `name` authenticated by
        `method`, within the caller's tenant, as a data URL. Readable by any
        authenticated user. Clients cache it by `(name, method, version)`,
        taking `version` from `GET /api/v1/people`.
      parameters:
        - name: method
          in: path
          required: true
          schema: {type: string}
        - name: name
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: The picture.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Avatar'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/people:
    get:
      tags: [auth]
      summary: Directory of the tenant's users
      description: |
        Lists the enabled users of the caller's tenant with just what the UI
        needs to render and pick people (owner avatars, the owner filter, the
        "Assign to…" picker). Readable by any authenticated user; carries no
        roles, groups or secrets.
      responses:
        '200':
          description: The directory, sorted by display name then login.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/Person'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT. The Python 1.x `Authorization: JWT <token>` scheme
        is rejected. Obtain a token via `POST /api/v1/login/local`.

    ingestToken:
      type: http
      scheme: bearer
      description: |
        Per-tenant ingest token. Opaque random string stored in the tenant
        document (`ingest_token` field). Supply as `Authorization: Bearer
        <ingest_token>` or `?token=<ingest_token>`. When absent or unknown,
        the middleware falls back to the `default` tenant (D4). This scheme
        gates `/api/v1/alerts` and `/api/v1/webhook/*` in multi-tenant
        deployments; single-tenant deployments leave ingest open (existing
        behaviour preserved).

    apiKeyAuth:
      type: http
      scheme: bearer
      description: >
        A user API key, presented as `Authorization: Bearer snz_…`. Keys are
        minted at `/api/v1/user/me/apikeys`, carry a subset of the owner's
        permissions, and are bounded live by the owner's current permissions on
        every request.

    proxyHeader:
      type: apiKey
      in: header
      name: X-Forwarded-User
      description: >
        Trusted-header (auth-proxy) mode. Gated by `auth_proxy.enabled` and
        **disabled by default**. When enabled, an upstream reverse proxy
        (oauth2-proxy / Pomerium / mod_auth_openidc) that has already
        authenticated the user forwards the identity in the configured username
        header (default `X-Forwarded-User`) and an optional groups header
        (default `X-Forwarded-Groups`). Snooze JIT-provisions a `method=proxy`
...
  responses:
...
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrEnvelope'
...
  schemas:
...
    Person:
      type: object
      required: [name, method]
      properties:
        name:
          type: string
          description: Login.
        method:
          type: string
          description: Auth method (`local`, `ldap`, `oidc`, …).
        display_name:
          type: string
        avatar_version:
          type: string
          description: |
            Content hash of the user's profile picture; empty or absent when
            the user has none (render initials instead).
    Avatar:
      type: object
      required: [name, method, version, data]
      properties:
        name: {type: string}
        method: {type: string}
        version: {type: string}
        data:
          type: string
          description: '`data:image/png;base64,…` URL.'
...
    ErrEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrBody'
    ErrBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - validation_error
            - unavailable
            - internal
        message: {type: string}
        details:
          type: object
          additionalProperties: true
        request_id: {type: string}
...